pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "promptval"
version = "0.1.0"
description = "Prompt Directory Validator: validate and optionally fix prompt .txt files for redundancy, conflicts, completeness, and PII."
readme = "README.md"
authors = [
  { name = "Pibit Candidate", email = "candidate@example.com" }
]
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "pydantic>=2.0,<3.0",
]
keywords = ["prompts", "validation", "cli", "pii", "quality"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://example.com/promptval"

[project.optional-dependencies]
openai = [
  "openai>=1.0.0,<2.0.0",
]
anthropic = [
  "anthropic>=0.34.0",
]
gemini = [
  "google-generativeai>=0.7.2",
]
all = [
  "openai>=1.0.0,<2.0.0",
  "anthropic>=0.34.0",
  "google-generativeai>=0.7.2",
]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
]

[project.scripts]
promptval = "promptval.cli:app"

[tool.pytest.ini_options]
addopts = "-q"

[tool.coverage.run]
source = ["promptval"]
branch = true